cambio de la cronjob
quitamos la cronjob por faltra de efectividad ahora todo se hace cuando acepta la compra

+++ database/migrations/2021_06_09_174047_add_fields_table_user.php
@@ -16,7 +16,7 @@ class AddFieldsTableUser extends Migration
         Schema::table('users', function (Blueprint $table) {
             $table->string('last_name')->after('name')->nullable();
             $table->bigInteger('master')->after('last_name')->nullable();
-            $table->integer('puntos')->after('master')->default(0);
+            $table->float('puntos')->after('master')->default(0);
             $table->integer('rango')->after('puntos')->default(0);
         });
     }
